chore: separate db package
