+++ .eslintignore
@@ -1,2 +1,2 @@
 node_modules
-packages/api/prisma/client
+packages/db/src/client
