+++ package.json
@@ -27,6 +27,7 @@
     "tslib": "2.3.0"
   },
   "devDependencies": {
+    "@babel/core": "7.14.5",
     "@babel/preset-react": "7.14.5",
     "@fortawesome/fontawesome-svg-core": "6.3.0",
     "@fortawesome/free-solid-svg-icons": "6.3.0",
@@ -34,16 +35,17 @@
     "@headlessui/react": "1.7.14",
     "@headlessui/tailwindcss": "0.1.3",
     "@nx/aws-lambda": "16.4.0",
-    "@nx/devkit": "16.3.2",
-    "@nx/eslint-plugin": "16.3.2",
-    "@nx/js": "16.3.2",
-    "@nx/linter": "16.3.2",
-    "@nx/next": "16.3.2",
-    "@nx/plugin": "16.3.2",
-    "@nx/react": "16.3.2",
-    "@nx/webpack": "16.3.2",
-    "@nx/workspace": "16.3.2",
+    "@nx/devkit": "16.7.1",
+    "@nx/eslint-plugin": "16.7.1",
+    "@nx/js": "16.7.1",
+    "@nx/linter": "16.7.1",
+    "@nx/next": "16.7.1",
+    "@nx/plugin": "16.7.1",
+    "@nx/react": "16.7.1",
+    "@nx/webpack": "16.7.1",
+    "@nx/workspace": "16.7.1",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.7",
+    "@prisma/nextjs-monorepo-workaround-plugin": "5.1.1",
     "@svgr/webpack": "8.0.1",
     "@swc-node/register": "1.4.2",
     "@swc/cli": "0.1.62",
@@ -52,15 +54,15 @@
     "@types/aws-lambda": "8.10.92",
     "@types/node": "18.11.9",
     "@types/nodemailer": "6.4.8",
-    "@types/react": "18.0.25",
-    "@types/react-dom": "18.0.9",
+    "@types/react": "18.2.14",
+    "@types/react-dom": "18.2.6",
     "@types/react-router-dom": "5.3.3",
     "@types/swagger-ui-react": "4.11.0",
-    "@typescript-eslint/eslint-plugin": "5.59.9",
-    "@typescript-eslint/parser": "5.59.9",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
     "autoprefixer": "10.4.13",
     "esbuild": "0.17.5",
-    "eslint": "8.15.0",
+    "eslint": "8.46.0",
     "eslint-config-next": "13.1.1",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-import": "2.26.0",
@@ -73,7 +75,7 @@
     "i18next-http-backend": "2.2.0",
     "jsonc-eslint-parser": "2.1.0",
     "lint-staged": "13.2.2",
-    "nx": "16.3.2",
+    "nx": "16.7.1",
     "postcss": "8.4.21",
     "prettier": "2.6.2",
     "react-hook-form": "7.43.9",
@@ -81,7 +83,7 @@
     "react-refresh": "0.10.0",
     "tailwindcss": "3.2.7",
     "ts-node": "10.9.1",
-    "typescript": "5.0.4",
+    "typescript": "5.1.6",
     "url-loader": "4.1.1",
     "zod": "3.21.4"
   },
