+++ packages/api/lucia.d.ts
@@ -3,6 +3,6 @@ declare namespace Lucia {
   type Auth = import('./shared/auth').Auth;
   type UserAttributes = {
     name?: string;
-    emailStatus?: import('.prisma/client').EmailStatus;
+    emailStatus?: import('@translation/db').EmailStatus;
   };
 }
