+++ packages/api/pages/api/languages/schemas.ts
@@ -3,6 +3,7 @@ import { Language } from '@translation/api-types';
 
 const schemaForType =
   <T>() =>
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   <S extends z.ZodType<T, any, any>>(arg: S) => {
     return arg;
   };
