+++ packages/db/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@translation/db",
+  "version": "0.0.1",
+  "type": "commonjs"
+}
