+++ packages/db/src/seed.ts
@@ -1,6 +1,6 @@
 import { bookKeys } from '../../../data/book-keys';
 import { morphologyData } from '../../../data/morphology';
-import { PrismaClient } from '../prisma/client';
+import { PrismaClient } from './client';
 
 const client = new PrismaClient();
